requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
frozenlist==1.3.3
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
openpyxl==3.1.2
pandas==2.0.1
pandas-stubs==2.0.1.230501
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
requests==2.29.0
six==1.16.0
soupsieve==2.4.1
textblob==0.17.1
tqdm==4.65.0
types-pytz==2023.3.0.0
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.3.3
yarl==1.9.2
